+++ src/redturtle/importer/base/interfaces.py
@@ -1,2 +1,8 @@
 # -*- coding: utf-8 -*-
-"""Module where all interfaces, events and exceptions live."""
+from zope.interface import Interface
+
+
+class IMigrationContextSteps(Interface):
+    """
+    Marker interface for specific context steps
+    """